user_service/requirements.txt
asgiref==3.9.2
Django==5.2.6
sqlparse==0.5.3
typing_extensions==4.15.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0
django-cors-headers==4.3.1
dj-database-url==2.1.0
drf-spectacular
django-allauth==65.3.0